backend/pyproject.toml
[project]
name = "substrack-backend"
version = "1.0.0"
description = "SubsTrack Backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "cryptography>=41.0.0",
    "aiosqlite==0.19.0",
    "httpx>=0.24.0",
    "tomlkit>=0.13.3",
    "asyncpg>=0.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
